pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ayase"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
    "sqlalchemy>=2.0.31",
    "alembic>=1.13.2",
    "python-dotenv>=1.0.1",
    "psycopg2>=2.9.9",
    "discord.py>=2.4.0",
    "pillow>=10.4.0",
]

[tool.setuptools.packages.find]
include = ["main.py"]
